+++ src/Transformers/RequestTransformer.php
@@ -27,11 +27,11 @@ class RequestTransformer implements TransformerInterface
      *
      * @param \Illuminate\Http\Request $data
      * @param int                      $limit
-     * @param int                      $time
+     * @param float                    $time
      *
      * @return \GrahamCampbell\Throttle\Data
      */
-    public function transform($data, int $limit = 10, int $time = 60)
+    public function transform($data, int $limit = 10, float $time = 60)
     {
         return new Data((string) $data->getClientIp(), (string) $data->path(), (int) $limit, (int) $time);
     }
